requirements.txt
asgiref==3.7.2
Django==4.2.4
django-browser-reload==1.11.0
django-cleanup==8.0.0
django-email-blacklist==1.0.1
django-mathfilters==1.0.0
Pillow==10.0.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
typing_extensions==4.7.1
python-dotenv==1.0.0
psycopg2==2.9.7
dj-database-url==2.1.0
gunicorn==21.0.1
whitenoise==6.5.0
six==1.16.0
boto3==1.34.0
django-storages==1.14.2
